Revert "[air][example] Dreambooth fine-tuning example fixes (#37504)"
This reverts commit 6ac4521f2a666b1dff8912f455ac928cbe91c5f4.
