+++ python/ray/air/examples/dreambooth/dataset.py
@@ -9,7 +9,7 @@ from transformers import AutoTokenizer
 
 def get_train_dataset(args, image_resolution=512):
     """Build a Dataset for fine-tuning DreamBooth model."""
-    # Load a directory of images as a Ray Dataset
+    # Load images into Dataset
     instance_dataset = read_images(args.instance_images_dir)
     class_dataset = read_images(args.class_images_dir)
 
@@ -17,18 +17,6 @@ def get_train_dataset(args, image_resolution=512):
     # two sets contain exactly the same number of images.
     # This is so we can zip them up during training to compute the
     # prior preserving loss in one pass.
-    #
-    # Example: If we have 200 class images (for regularization) and 4 instance
-    # images of our subject, then we'll duplicate the instance images 50 times
-    # so that our dataset looks like:
-    #
-    #     instance_image_0, class_image_0
-    #     instance_image_1, class_image_1
-    #     instance_image_2, class_image_2
-    #     instance_image_3, class_image_3
-    #     instance_image_0, class_image_4
-    #     instance_image_1, class_image_5
-    #     ...
     dup_times = class_dataset.count() // instance_dataset.count()
     instance_dataset = instance_dataset.map_batches(
         lambda df: pd.concat([df] * dup_times), batch_format="pandas"
@@ -53,97 +41,65 @@ def get_train_dataset(args, image_resolution=512):
     class_prompt_ids = _tokenize(args.class_prompt)[0]
     instance_prompt_ids = _tokenize(args.instance_prompt)[0]
 
-    # START: image preprocessing
+    # Image preprocessing.
+    # Instance and class images used by this example are in sizes 700x700
+    # and 512x512 respectively.
+    # Depending on the sizes of actual training images, there may need to be a
+    # transforms.Resize() step as well.
     transform = transforms.Compose(
         [
             transforms.ToTensor(),
-            transforms.Resize(
-                image_resolution,
-                interpolation=transforms.InterpolationMode.BILINEAR,
-                antialias=True,
-            ),
             transforms.RandomCrop(image_resolution),
             transforms.Normalize([0.5], [0.5]),
         ]
     )
-    instance_ds_preprocessor = TorchVisionPreprocessor(
-        columns=["image"], output_columns=["instance_image"], transform=transform
-    )
-    class_ds_preprocessor = TorchVisionPreprocessor(
-        columns=["image"], output_columns=["class_image"], transform=transform
-    )
-    # END: image preprocessing
+    preprocessor = TorchVisionPreprocessor(["image"], transform=transform)
 
-    # START: Apply preprocessing steps as Ray Dataset operations
-    # For each dataset:
-    # - perform image preprocessing
-    # - drop the original image column
-    # - add a new column with the tokenized prompts
-    instance_dataset = (
-        instance_ds_preprocessor.transform(instance_dataset)
-        .drop_columns(["image"])
-        .add_column("instance_prompt_ids", lambda df: [instance_prompt_ids] * len(df))
+    instance_dataset = preprocessor.transform(instance_dataset).add_column(
+        "prompt_ids", lambda df: [instance_prompt_ids] * len(df)
     )
-    # END: Apply preprocessing steps as Ray Dataset operations
-
-    class_dataset = (
-        class_ds_preprocessor.transform(class_dataset)
-        .drop_columns(["image"])
-        .add_column("class_prompt_ids", lambda df: [class_prompt_ids] * len(df))
+    class_dataset = preprocessor.transform(class_dataset).add_column(
+        "prompt_ids", lambda df: [class_prompt_ids] * len(df)
     )
-    # --- Ray Data
-
-    # We may have too many duplicates of the instance images, so limit the
-    # dataset size so that len(instance_dataset) == len(class_dataset)
-    final_size = min(instance_dataset.count(), class_dataset.count())
+    # ---
 
     # Now, zip the images up.
+    final_size = min(instance_dataset.count(), class_dataset.count())
     train_dataset = (
         instance_dataset.limit(final_size)
         .repartition(final_size)
         .zip(class_dataset.limit(final_size).repartition(final_size))
     )
 
-    print("Training dataset schema after pre-processing:")
-    print(train_dataset.schema())
-
     return train_dataset.random_shuffle()
 
 
 def collate(batch, device, dtype):
-    """Build Torch training batch.
+    """Build Torch training batch."""
+    # Layout of the batch is that instance image data (pixels, prompt ids) occupy
+    # the top half of the batch. And class image data occupy the bottom half
+    # of the batch.
+    # During training, a batch will be chunked into 2 sub-batches for prior
+    # preserving loss calculation.
 
-    B = batch size
-    (C, W, H) = (channels, width, height)
-    L = max length in tokens of the text guidance input
+    # batch["image"] = image1, image2
+    # batch["image_1"] = reg1, reg2
+    # After cat, we will have [image1, reg1, image2, reg]
 
-    Input batch schema (see `get_train_dataset` on how this was setup):
-        instance_images: (B, C, W, H)
-        class_images: (B, C, W, H)
-        instance_prompt_ids: (B, L)
-        class_prompt_ids: (B, L)
-
-    Output batch schema:
-        images: (2 * B, C, W, H)
-            All instance images in the batch come before the class images:
-            [instance_images[0], ..., instance_images[B-1], class_images[0], ...]
-        prompt_ids: (2 * B, L)
-            Prompt IDs are ordered the same way as the images.
-
-    During training, a batch will be chunked into 2 sub-batches for
-    prior preserving loss calculation.
-    """
-
-    images = torch.cat([batch["instance_image"], batch["class_image"]], dim=0)
+    images = torch.cat([batch["image"], batch["image_1"]], dim=0)
     images = images.to(memory_format=torch.contiguous_format).float()
 
-    batch_size = len(batch["instance_prompt_ids"])
+    batch_size = len(batch["prompt_ids"])
 
-    prompt_ids = torch.cat(
-        [batch["instance_prompt_ids"], batch["class_prompt_ids"]], dim=0
+    # batch["prompt_ids"] = pr1, pr2
+    # batch["prompt_ids_1"] = rr1, rr2
+    # After stack+reshape, we will have [pr1, rr1, pr2, rr2]
+
+    prompt_ids = torch.stack(
+        [batch["prompt_ids"], batch["prompt_ids_1"]], dim=1
     ).reshape(batch_size * 2, -1)
 
     return {
-        "images": images.to(device, dtype=dtype),
         "prompt_ids": prompt_ids.to(device),  # token ids should stay int.
+        "images": images.to(device, dtype=dtype),
     }
